pyproject.toml
[tool.poetry]
name = "icon"
version = "0.1.0"
description = "Script to generate my icon"
authors = ["keke1008 <keke2222337@gmail.com>"]
readme = "README.md"
classifiers = ["Private :: Do not Upload"]

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
